Add JVM class file decompiler: .class/.jar to Java source
- Java class file parser (JVMS ch. 4): constant pool, members, Code +
  exception table, ConstantValue, Exceptions, InnerClasses, Signature,
  LocalVariableTable, BootstrapMethods
- JVM opcode table (JVMS ch. 6) + bytecode decoder with wide/switch
  support and absolute branch targets
- Java source decompiler: classes/interfaces/enums, javac-shape loops,
  for/while/do-while, switch inlining, ternary + boolean-return
  reconstruction, try/catch/finally, foreach from the iterator pattern,
  synchronized blocks, invokedynamic string concat recipes
- Generic Signature rendering: type params, parameterized types,
  wildcards in declarations
- .jar support: zip reader + built-in raw-DEFLATE decoder (no deps)
- javap-style disassembler (--il on .class/.jar)
- Shared control-flow engine (decompile/flow.rs) used by both back ends;
  fixes latent binop parenthesis bug in the C# path
- CLI: magic-byte dispatch (MZ / 0xCAFEBABE / PK), --list/--il/--type/
  --stdout/--recursive across all formats
- Doc-drift enforcement (tests/docs_sync.rs) and process updates
  ported from mathr: CRAP score, codebase scorecard, mature module policy
- Fixtures: Java sources + prebuilt classes and jars; decompiled output
  verified to recompile with javac

+++ src/metadata/reader.rs
@@ -293,9 +293,9 @@ impl<'a> Reader<'a> {
                     let mut owner_name = String::new();
                     let type_defs = self.tables.get(tbl::TYPEDEF);
                     for (i, td) in type_defs.iter().enumerate() {
-                        let start = td.col(5) as u32;
-                        let next = type_defs.get(i + 1).map(|r| r.col(5) as u32)
-                            .unwrap_or_else(|| self.tables.row_count(tbl::METHODDEF) as u32 + 1);
+                        let start = td.col(5);
+                        let next = type_defs.get(i + 1).map(|r| r.col(5))
+                            .unwrap_or_else(|| self.tables.row_count(tbl::METHODDEF) + 1);
                         if decl_ci.row >= start && decl_ci.row < next {
                             owner_name = self.type_def_name(td);
                             break;
@@ -421,7 +421,7 @@ impl<'a> Reader<'a> {
                 }
                 let (len, len_bytes) = crate::metadata::streams::decode_compressed_uint(&val[*pos..]).ok()?;
                 *pos += len_bytes;
-                let end = *pos + len as usize;
+                let end = *pos + len;
                 if end > val.len() { return None; }
                 let s = std::str::from_utf8(&val[*pos..end]).ok()?;
                 *pos = end;
@@ -462,9 +462,9 @@ impl<'a> Reader<'a> {
                 let m_row = ctor.row;
                 let type_defs = self.tables.get(tbl::TYPEDEF);
                 for (i, td) in type_defs.iter().enumerate() {
-                    let start = td.col(5) as u32;
-                    let next = type_defs.get(i + 1).map(|r| r.col(5) as u32)
-                        .unwrap_or_else(|| self.tables.row_count(tbl::METHODDEF) as u32 + 1);
+                    let start = td.col(5);
+                    let next = type_defs.get(i + 1).map(|r| r.col(5))
+                        .unwrap_or_else(|| self.tables.row_count(tbl::METHODDEF) + 1);
                     if m_row >= start && m_row < next {
                         let name = self.type_def_name(td);
                         return Some(name);
